requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
marshmallow==3.19.0
marshmallow-enum==1.5.1
jsonschema==4.17.3
pendulum==2.1.2
werkzeug==2.2.3
orjson==3.8.7
loguru==0.6.0
gunicorn==20.1.0
MarkupSafe==2.1.1
numpy==1.24.2
openai==0.27.0
pandas==1.5.3
twilio==8.1.0
tiktoken==0.3.3
pinecone-client==2.2.1
psycopg2-binary
mypy
uuid
